move cartrepr_from_matmul to utils
